added home page
closes #5
