+++ lib/custom_mesages.dart
@@ -1,4 +1,3 @@
 class CustomMessages {
   static String signInText = "Sign In";
-
 }
